requirements.txt
glaip-sdk==0.0.2
python-dotenv
